+++ src/main/kotlin/graphql/dataTransferObjects/GroupChatInviteMessage.kt
@@ -1,12 +1,12 @@
 package com.neelkamath.omniChatBackend.graphql.dataTransferObjects
 
+import com.neelkamath.omniChatBackend.db.tables.GroupChatInviteMessages
 import com.neelkamath.omniChatBackend.db.tables.GroupChats
-import com.neelkamath.omniChatBackend.db.tables.Messages
 import java.util.*
 
 class GroupChatInviteMessage(override val id: Int) : Message, ReadMessageResult {
     fun getInviteCode(): UUID? {
-        val chatId = Messages.readChatId(id)
+        val chatId = GroupChatInviteMessages.read(id)
         return GroupChats.readInviteCode(chatId)
     }
 }