+++ docs/CHANGELOG.md
@@ -58,6 +58,7 @@ The entire project (i.e., the GraphQL API, REST API, and server) uses the same v
 ### Fixed
 
 - Fix `Mutation.deleteAccount` so that it doesn't crash when the user has messages in chats they're no longer in.
+- Fix the `inviteCode` field for `type GroupChatInviteMessage`, `type NewGroupChatInviteMessage`, and `type StarredGroupChatInviteMessage`.
 
 ## [0.21.0](https://github.com/neelkamath/omni-chat-backend/releases/tag/v0.21.0) - 2021-06-16